migrate wrapper JSX to HOC
