+++ packages/charlesstover.com/src/features/home/home.hoc.ts
@@ -1,0 +1,5 @@
+import withWrapper from '../../hocs/with-wrapper';
+import Home from './home.view';
+import useWrapperProps from './hooks/use-wrapper-props';
+
+export default withWrapper(Home, useWrapperProps);
