+++ packages/charlesstover.com/src/features/home/hooks/use-wrapper-props.ts
@@ -1,0 +1,10 @@
+import type { Props as WrapperProps } from '../../../components/wrapper';
+
+export default function useHomeWrapperProps(): Omit<
+  WrapperProps,
+  'breadcrumbs' | 'children'
+> {
+  return {
+    toolsHide: true,
+  };
+}
