+++ packages/charlesstover.com/src/features/home/index.ts
@@ -1,1 +1,1 @@
-export { default } from './home.view';
+export { default } from './home.hoc';
