--- packages/charlesstover.com/src/features/packages/components/content/index.ts
@@ -1,1 +1,0 @@
-export { default } from './content.view';
