+++ packages/charlesstover.com/src/features/publications/components/banner/banner.module.scss
@@ -1,11 +1,12 @@
 @mixin chip($x, $y) {
   $border-color: #404040;
-  background-color: rgba(64 64 64 / 85%) !important;
+  background-color: rgba(64 64 64 / 90%) !important;
   border-bottom-color: $border-color !important;
   border-left-color: $border-color !important;
   border-top-color: $border-color !important;
   border-style: solid !important;
   border-width: 0 !important;
+  box-shadow: 0.125em 0.125em 0.25em 0.25em #000000;
   line-height: 1em;
   position: absolute;
 
@@ -46,6 +47,7 @@
   line-height: 0;
   margin: 0 auto;
   opacity: 0.8;
+  overflow: hidden;
   position: relative;
   width: fit-content;
