+++ packages/charlesstover.com/src/features/publications/components/banner/banner.view.tsx
@@ -3,7 +3,7 @@ import NumberFormat from 'number-format-react';
 import type { ReactElement } from 'react';
 import Chip from '../../../../components/chip';
 import Link from '../../../../components/link';
-import Quantity from '../../../../components/quantity';
+// import Quantity from '../../../../components/quantity';
 import validateString from '../../../../utils/validate-string';
 import BannerImage from '../../components/banner-image';
 import styles from './banner.module.scss';
@@ -20,33 +20,27 @@ interface Props {
 
 const bannerHeightClassName: string = validateString(styles.bannerHeight);
 const bannerWidthClassName: string = validateString(styles.bannerWidth);
-const BASE = 10;
-const dateTimeClassName: string = validateString(styles.dateTime);
+// const dateTimeClassName: string = validateString(styles.dateTime);
 const linkClassName: string = validateString(styles.link);
-const MONTH_OFFSET = 1;
+// const MONTH_OFFSET = 1;
 const reactionsClassName: string = validateString(styles.reactions);
-const readingTimeClassName: string = validateString(styles.readingTime);
+// const readingTimeClassName: string = validateString(styles.readingTime);
 const viewsClassName: string = validateString(styles.views);
 
-const prefix = (n: number): string => {
-  if (n < BASE) {
-    return `0${n}`;
-  }
-  return n.toString();
-};
-
 export default function PublicationBanner({
-  dateTime,
+  // dateTime,
   image,
   reactions,
-  readingTime,
+  // readingTime,
   title,
   url,
   views,
 }: Readonly<Props>): ReactElement {
+  /*
   const date: Date = new Date(dateTime);
   const monthStr: string = prefix(date.getMonth() + MONTH_OFFSET);
   const dayStr: string = prefix(date.getDate());
+  */
   return (
     <div className={bannerHeightClassName}>
       <div className={bannerWidthClassName}>
@@ -64,6 +58,7 @@ export default function PublicationBanner({
         <Chip className={reactionsClassName} title={<I18n>Reactions</I18n>}>
           <NumberFormat>{reactions}</NumberFormat> 👏
         </Chip>
+        {/*
         <Chip
           className={dateTimeClassName}
           title={<I18n>Publication date</I18n>}
@@ -78,6 +73,7 @@ export default function PublicationBanner({
             <Quantity unit="minutes">{readingTime}</Quantity> ⏳
           </Chip>
         )}
+        */}
       </div>
     </div>
   );
