+++ packages/charlesstover.com/src/features/publications/components/stats/stats.module.scss
@@ -3,14 +3,14 @@
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
-  justify-content: space-between;
+  // justify-content: space-between;
   margin-bottom: -1em;
   margin-left: -1em;
   margin-right: -1em;
 
   > div {
     display: flex;
-    flex-basis: 0;
+    flex-grow: 1;
     flex-direction: column;
     margin-bottom: 1em;
     margin-left: 1em;
