+++ packages/charlesstover.com/src/features/publications/components/stats/stats.view.tsx
@@ -2,8 +2,8 @@ import I18n from 'lazy-i18n';
 import type { ReactElement } from 'react';
 import Span from '../../../../components/span';
 import mapTimeToDaysAgo from '../../../../utils/map-time-to-days-ago';
+import round from '../../../../utils/round';
 import validateString from '../../../../utils/validate-string';
-import ratio from '../../utils/ratio';
 import styles from './stats.module.scss';
 
 interface Props {
@@ -12,33 +12,41 @@ interface Props {
   readonly views: number;
 }
 
+const DECIMALS = 2;
+const LOCKED_DAY = 22;
+const LOCKED_MONTH = 6; // July
+const LOCKED_YEAR = 2023;
+const PERCENT = 100;
 const rootClassName: string = validateString(styles.root);
-const TWO = 2;
+const LOCKED_DATE: Date = new Date(LOCKED_YEAR, LOCKED_MONTH, LOCKED_DAY);
+const LOCKED_TIME: number = LOCKED_DATE.getTime();
+const LOCKED_DAYS_AGO: number = mapTimeToDaysAgo(LOCKED_TIME);
 
 export default function PublicationStats({
   dateTime,
   reactions,
   views,
 }: Readonly<Props>): ReactElement {
+  const days: number = mapTimeToDaysAgo(dateTime) - LOCKED_DAYS_AGO;
   return (
     <div className={rootClassName}>
       <div>
         <Span color="label" size="small-heading">
-          <I18n>Reactions/day</I18n>
+          <I18n>Daily views</I18n>
         </Span>
-        <Span>{ratio(reactions, mapTimeToDaysAgo(dateTime))}</Span>
+        <Span>{round(views / days, DECIMALS)}</Span>
       </div>
       <div>
         <Span color="label" size="small-heading">
-          <I18n>Reactions/view</I18n>
+          <I18n>Daily reactions</I18n>
         </Span>
-        <Span>{ratio(reactions, views, TWO)}%</Span>
+        <Span>{round(reactions / days, DECIMALS)}</Span>
       </div>
       <div>
         <Span color="label" size="small-heading">
-          <I18n>Views/day</I18n>
+          <I18n>Engagement rate</I18n>
         </Span>
-        <Span>{ratio(views, mapTimeToDaysAgo(dateTime))}</Span>
+        <Span>{round((reactions / views) * PERCENT, DECIMALS)}%</Span>
       </div>
     </div>
   );
