+++ packages/charlesstover.com/src/features/publications/hooks/use-wrapper-props.tsx
@@ -1,10 +1,10 @@
 import type { TranslateFunction } from 'lazy-i18n';
 import I18n, { useTranslate } from 'lazy-i18n';
 import { useMemo } from 'react';
-import type { Props as WrapperProps } from '../../../components/wrapper';
 import type Breadcrumb from '../../../types/breadcrumb';
+import type { Props as WrapperProps } from '../../../components/wrapper';
 
-export default function useDashboardWrapperProps(): Omit<
+export default function usePublicationsWrapperProps(): Omit<
   WrapperProps,
   'children'
 > {
@@ -13,14 +13,14 @@ export default function useDashboardWrapperProps(): Omit<
 
   // States
   return {
-    fallback: <I18n>Loading dashboard</I18n>,
+    fallback: <I18n>Loading publications</I18n>,
     toolsHide: true,
 
     breadcrumbs: useMemo(
       (): readonly Breadcrumb[] => [
         {
-          children: translate('Dashboard') ?? '...',
-          path: '/dashboard',
+          children: translate('Publications') ?? '...',
+          path: '/publications',
         },
       ],
       [translate],
