+++ packages/charlesstover.com/src/features/publications/publications.hook.ts
@@ -1,24 +1,95 @@
 import type { TranslateFunction } from 'lazy-i18n';
 import { useTranslate } from 'lazy-i18n';
-import { useMemo } from 'react';
-import type Breadcrumb from '../../types/breadcrumb';
+import { useCallback, useMemo, useState } from 'react';
+import useDevStats from '../../hooks/use-dev-stats';
+import useMediumStats from '../../hooks/use-medium-stats';
+import Sort from './constants/publications-sort';
+import useItems from './hooks/use-content-items';
+import type Publication from './types/publication';
+import filterByPublicationsSort from './utils/filter-by-publications-sort';
+import filterItemsByMinimumViews from './utils/filter-publications-items-by-minimum-views';
+import mapSortToFunction from './utils/map-publications-sort-to-function';
 
 interface State {
-  readonly breadcrumbs: readonly Breadcrumb[];
+  readonly handleMediumApiBannerDismiss: VoidFunction;
+  readonly handleMinimumViewsBannerDismiss: VoidFunction;
+  readonly handleSortChange: (sort: string | undefined) => void;
+  readonly isMediumApiBannerVisible: boolean;
+  readonly isMinimumViewsBannerVisible: boolean;
+  readonly items: readonly Publication[];
+  readonly loading: string | undefined;
+  readonly sort: Sort;
 }
 
-export default function usePublications(): State {
+const DEFAULT_SORT: Sort = Sort.ViewsPerDay;
+
+export default function usePublicationsContents(): State {
+  // Contexts
   const translate: TranslateFunction = useTranslate();
 
+  const {
+    data: devData,
+    // error: devError,
+    isLoading: isDevLoading,
+    // refetch: refetchDev,
+  } = useDevStats();
+
+  const {
+    data: mediumData,
+    // error: mediumError,
+    isLoading: isMediumLoading,
+    // refetch: refetchMedium,
+  } = useMediumStats();
+
+  // States
+  const [sort, setSort] = useState<Sort>(DEFAULT_SORT);
+  const [isMediumApiBannerVisible, setIsMediumApiBannerVisible] =
+    useState(true);
+  const [isMinimumViewsBannerVisible, setIsMinimumViewsBannerVisible] =
+    useState(true);
+
+  const items: readonly Publication[] = useItems({
+    devData,
+    mediumData,
+  });
+
   return {
-    breadcrumbs: useMemo(
-      (): readonly Breadcrumb[] => [
-        {
-          children: translate('Publications') ?? '...',
-          path: '/publications',
-        },
-      ],
-      [translate],
-    ),
+    isMediumApiBannerVisible,
+    isMinimumViewsBannerVisible,
+    sort,
+
+    handleMediumApiBannerDismiss: useCallback((): void => {
+      setIsMediumApiBannerVisible(false);
+    }, []),
+
+    handleMinimumViewsBannerDismiss: useCallback((): void => {
+      setIsMinimumViewsBannerVisible(false);
+    }, []),
+
+    handleSortChange: useCallback((newSort: string | undefined): void => {
+      if (typeof newSort === 'undefined') {
+        setSort(DEFAULT_SORT);
+        return;
+      }
+      if (!filterByPublicationsSort(newSort)) {
+        throw new Error(
+          `Expected a publication sort option, but received: ${newSort}`,
+        );
+      }
+      setSort(newSort);
+    }, []),
+
+    items: useMemo((): readonly Publication[] => {
+      const newItems: Publication[] = [...items];
+      newItems.sort(mapSortToFunction(sort));
+      return newItems.filter(filterItemsByMinimumViews);
+    }, [items, sort]),
+
+    loading: useMemo((): string | undefined => {
+      if (!isDevLoading && !isMediumLoading) {
+        return;
+      }
+      return translate('Loading publications') ?? '...';
+    }, [isDevLoading, isMediumLoading, translate]),
   };
 }
