--- packages/charlesstover.com/src/features/quotes/components/content/content.view.tsx
@@ -1,16 +1,0 @@
-import type { ReactElement } from 'react';
-import Cards from '../../../../components/cards';
-import CardContent from '../../components/card-content';
-import CardHeader from '../../components/card-header';
-import ITEMS from '../../constants/quotes';
-
-export default function QuotesContent(): ReactElement {
-  return (
-    <Cards
-      CardContent={CardContent}
-      CardHeader={CardHeader}
-      cardKey="author"
-      items={ITEMS}
-    />
-  );
-}
