+++ packages/charlesstover.com/src/features/quotes/index.ts
@@ -1,1 +1,1 @@
-export { default } from './quotes.view';
+export { default } from './quotes.hoc';
