+++ packages/charlesstover.com/src/features/quotes/quotes.hoc.tsx
@@ -1,0 +1,8 @@
+import { lazy } from 'react';
+import withWrapper from '../../hocs/with-wrapper';
+import useWrapperProps from './hooks/use-wrapper-props';
+
+export default withWrapper(
+  lazy(async () => import('./quotes.view')),
+  useWrapperProps,
+);
