+++ packages/charlesstover.com/src/features/quotes/quotes.view.tsx
@@ -1,23 +1,16 @@
-import I18n from 'lazy-i18n';
-import type { ComponentType, ReactElement } from 'react';
-import { lazy } from 'react';
-import Wrapper from '../../components/wrapper';
-import useQuotes from './quotes.hook';
-
-const Content: ComponentType<unknown> = lazy(
-  async () => import('./components/content'),
-);
+import type { ReactElement } from 'react';
+import Cards from '../../components/cards';
+import CardContent from './components/card-content';
+import CardHeader from './components/card-header';
+import ITEMS from './constants/quotes';
 
 export default function Quotes(): ReactElement {
-  const { breadcrumbs } = useQuotes();
-
   return (
-    <Wrapper
-      breadcrumbs={breadcrumbs}
-      fallback={<I18n>Loading quotes</I18n>}
-      toolsHide
-    >
-      <Content />
-    </Wrapper>
+    <Cards
+      CardContent={CardContent}
+      CardHeader={CardHeader}
+      cardKey="author"
+      items={ITEMS}
+    />
   );
 }
