+++ packages/charlesstover.com/src/translations/arabic.json
@@ -5,21 +5,19 @@
   "1 match": " تطابق ١",
   "1 minute": "دقيقه ١",
   "About me": "عني",
-  "An error occurred.": "An error occurred.",
   "Applications": "تطبيقات",
   "Avatar": "الصورة الرمزية",
   "Breadcrumbs": "فتات الخبز",
-  "breadcrumb": "breadcrumb",
   "Cancel": "إلغاء",
   "Close navigation": "غلق التنقل",
   "Close tools": "غلق الادوات",
   "Confirm": "تأكيد",
   "Connect with me": "تواصل معي",
+  "Daily reactions": "تفاعلات في اليوم",
+  "Daily views": "المشاهدات في اليوم",
   "Dark mode": "الوضع المظلم",
-  "Design system": "Design system",
-  "Direct downloads": "Direct downloads",
   "Dismiss": "صرف",
-  "Enterprise: $n years": "Enterprise: $n years",
+  "Engagement rate": "تفاعلات لكل مشاهدة",
   "external": "رابط خارجي",
   "Filter packages": "تنقيه الباقات",
   "Front end: $n years": "سنين $n :مطور الواجهة الأمامية",
@@ -27,49 +25,33 @@
   "Home": "الصفحة الرئيسية",
   "Loading packages": "تحميل الباقات",
   "Loading publications": "تحميل المنشورات",
-  "Loading quotes": "Loading quotes",
   "Loading spritesheet to GIF converter": "تحميل محول السبرايت شييت لالجرافيكية المتحركة",
-  "Main navigation": "Main navigation",
   "More applications": "تطبيقات اخري",
   "Navigation": "تنقل",
   "No matches": "لا مطابقات",
   "Notifications": "إشعارات",
-  "Only publications with more than $count views are shown.": "Only publications with more than $count views are shown.",
   "Package name": "أسم الباقه",
   "Packages": "الباقات",
   "Portfolio": "الملف",
   "Preferences": "التفضيلات",
-  "Projects": "Projects",
   "Properties": "خصائص",
   "Publication date": "تاريخ النشر",
   "Publications": "المنشورات",
   "Quotes": "أقتباسات",
   "Reactions": "تفاعلات",
-  "Reactions/day": "تفاعلات في اليوم",
-  "Reactions/view": "تفاعلات لكل مشاهدة",
-  "Reactions per day": "تفاعلات في اليوم",
-  "Reactions per view": "تفاعلات لكل مشاهدة",
   "Reading time": "وقت القراءة",
-  "Result": "Result",
-  "Retry": "Retry",
   "RPG overworld engine": "محرك الار بي جي اوفرورلد",
   "Select page size.": ".حدد حجم الصفحة",
   "Select to see all the text and wrap the lines.": "حدد لالتفاف النص وكشف كل النص",
   "Select visible columns.": "حدد الأعمدة المرئية",
   "Settings": "إعدادات",
-  "Show path": "Show path",
   "Sort by": "ترتيب حسب",
-  "sorted ascending": "sorted ascending",
-  "sorted descending": "sorted descending",
   "Sprite sheet to GIF": "سبرايت شييت لالجرافيكية المتحركة",
   "Toggle navigation": "تبديل التنقل",
   "Toggle tools": "تبديل الأدوات",
   "Tools": "أدوات",
   "Total downloads": "إجمالي التنزيلات",
-  "version: $version": "version: $version",
   "View résumé": "عرض السيرة الذاتية",
   "Views": "المشاهدات",
-  "Views/day": "المشاهدات في اليوم",
-  "Views per day": "المشاهدات في اليوم",
   "Wrap lines": "التفاف النص"
 }
