+++ packages/charlesstover.com/src/translations/cebuano.json
@@ -15,10 +15,13 @@
   "Close tools": "Isira ang mga gamit",
   "Confirm": "Ipagmatuod",
   "Connect with me": "Pakigsulti kanako",
+  "Daily reactions": "Mga reaksyon/adlaw",
+  "Daily views": "Mga lantaw/adlaw",
   "Dark mode": "Dark mode",
   "Design system": "Design system",
   "Direct downloads": "Deretso downloads",
   "Dismiss": "Tangtanga",
+  "Engagement rate": "Mga reaksyon/lantaw",
   "Enterprise: $n years": "Negosyo: $n ka tuig",
   "external": "gawas",
   "Filter packages": "Isala ang mga putos",
@@ -45,10 +48,6 @@
   "Publications": "Mga publikasyon",
   "Quotes": "Mga kinutlo",
   "Reactions": "Mga reaksyon",
-  "Reactions/day": "Mga reaksyon/adlaw",
-  "Reactions/view": "Mga reaksyon/lantaw",
-  "Reactions per day": "Mga reaksyon kada adlaw",
-  "Reactions per view": "Mga reaksyon kada lantaw",
   "Reading time": "Oras sa pagbasa",
   "Result": "Resulta",
   "Retry": "Suwaya'g usba",
@@ -69,7 +68,5 @@
   "version: $version": "bersyon: $version",
   "View résumé": "Tan-awa ang résumé",
   "Views": "Mga lantaw",
-  "Views/day": "Mga lantaw/adlaw",
-  "Views per day": "Mga lantaw kada adlaw",
   "Wrap lines": "Iputos ang mga linya"
 }
