+++ packages/charlesstover.com/src/translations/english.json
@@ -52,13 +52,17 @@
   "Cumulative Layout Shift": "Cumulative Layout Shift",
   "Currently online: $status": "Currently online: $status",
   "Customize the columns' visibility and order.": "Customize the columns' visibility and order.",
+  "Daily reactions": "Daily reactions",
+  "Daily views": "Daily views",
   "Dark mode": "Dark mode",
   "Dashboard": "Dashboard",
   "Dec.": "Dec.",
   "Design system": "Design system",
   "Direct downloads": "Direct downloads",
   "Dismiss": "Dismiss",
+  "Due to a dependency API change, publication statistics are locked to July 2023.": "Due to a dependency API change, publication statistics are locked to July 2023.",
   "Duration": "Duration",
+  "Engagement rate": "Engagement rate",
   "Enterprise: $n years": "Enterprise: $n years",
   "Error rate": "Error rate",
   "Errors": "Errors",
@@ -124,10 +128,6 @@
   "Publications": "Publications",
   "Quotes": "Quotes",
   "Reactions": "Reactions",
-  "Reactions/day": "Reactions/day",
-  "Reactions/view": "Reactions/view",
-  "Reactions per day": "Reactions per day",
-  "Reactions per view": "Reactions per view",
   "Reading time": "Reading time",
   "Request time": "Request time",
   "Response body size": "Response body size",
@@ -160,8 +160,6 @@
   "version: $version": "version: $version",
   "View résumé": "View résumé",
   "Views": "Views",
-  "Views/day": "Views/day",
-  "Views per day": "Views per day",
   "Wall time": "Wall time",
   "Web Vitals": "Web Vitals",
   "Wrap lines": "Wrap lines"
